add option to initialse make
The command 'make init' will touch all files (this will update their timestamps) and remove the main target files. Another run of 'make' will create the targets. This approach is useful when checking out the git repo for the first time and the only aim is to be able to perform the last step of the analysis (i.e. create the pdf report).